backend/requirements 2.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
langchain-openai>=0.0.2
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.14
sentence-transformers>=2.2.2
python-multipart>=0.0.6
pytest>=7.4.3
httpx>=0.25.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4 